feat: 과거 예약 삭제 검증 커스텀 예외 추가 및 적용

+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -6,6 +6,7 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.global.exception.policy.PastReservationCancelNotAllowedException;
 import roomescape.global.exception.policy.ReservationConflictException;
 import roomescape.global.exception.policy.PastReservationNotAllowedException;
 import roomescape.global.exception.validation.ThemeNotFoundException;
@@ -47,16 +48,13 @@ public class ReservationService {
         validateThemeExists(themeId);
         LocalDateTime reservationDateTime = LocalDateTime.of(date,
                 reservationTimeRepository.findById(timeId).getStartAt());
-        validateDateTimeIsNotPast(reservationDateTime);
+
+        if(isBeforeNow(reservationDateTime)) {
+            throw new PastReservationNotAllowedException();
+        }
         if (reservationRepository.existsByDateAndTimeIdAndThemeId(
                 new DuplicateReservationCondition(date, timeId, themeId))) {
             throw new ReservationConflictException();
-        }
-    }
-
-    private void validateDateTimeIsNotPast(LocalDateTime reservationDateTime) {
-        if (reservationDateTime.isBefore(LocalDateTime.now())) {
-            throw new PastReservationNotAllowedException();
         }
     }
 
@@ -70,8 +68,14 @@ public class ReservationService {
     @Transactional
     public void cancelReservation(Long id) {
         Reservation reservation = reservationRepository.findById(id);
-        validateDateTimeIsNotPast(reservation.getReservationDateTime());
+        if(isBeforeNow(reservation.getReservationDateTime())) {
+            throw new PastReservationCancelNotAllowedException();
+        }
 
         reservationRepository.deleteById(id);
+    }
+
+    private boolean isBeforeNow(LocalDateTime reservationDateTime) {
+        return reservationDateTime.isBefore(LocalDateTime.now());
     }
 }
